+++ pin-project-internal/src/pin_project/derive.rs
@@ -944,6 +944,7 @@ fn make_proj_impl(
 
     let mut project = Some(quote! {
         #allow_dead_code
+        #[inline]
         #vis fn project<#lifetime>(
             self: _pin_project::__private::Pin<&#lifetime mut Self>,
         ) -> #proj_ident #proj_ty_generics {
@@ -955,6 +956,7 @@ fn make_proj_impl(
     let mut project_ref = Some(quote! {
         #allow_dead_code
         #[allow(clippy::missing_const_for_fn)]
+        #[inline]
         #vis fn project_ref<#lifetime>(
             self: _pin_project::__private::Pin<&#lifetime Self>,
         ) -> #proj_ref_ident #proj_ty_generics {
@@ -967,6 +969,7 @@ fn make_proj_impl(
         // It is enough to only set the span of the signature.
         let sig = quote_spanned! { span =>
             #allow_dead_code
+            #[inline]
             #vis fn project_replace(
                 self: _pin_project::__private::Pin<&mut Self>,
                 __replacement: Self,